more refactoring- broke up APIClient and DataManager, more Result
